IncSearch and Search columns are the same
When using the incsearch option I've noticed that all the matches have a visible yellow background, apart from the first match which doesn't. As you type more characters into the search and narrow down the options the first match might move. When I look at one of the matches while I type and it becomes the first match, the disappearance of the yellow background is so immediate that it feels like I typed a wrong character and it's no longer a match.

Not sure this is an intentional design or not. If not, I believe this change is the fix. Having tried it out locally, it makes all the search matches look the same.